% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnWeight.R
\name{fnWeight}
\alias{fnWeight}
\title{A Flexibily Function to Calculate Proportions}
\usage{
fnWeight(dat, resp = NULL, term = NULL, calc = NULL)
}
\arguments{
\item{dat}{A dataset with a specified structure (details coming soon...)}

\item{resp}{The variable we would like to calculate a share for}

\item{term}{Any groupings we'd like to apply in our calculation}

\item{calc}{An r expression that will determine the numerator of our share calculation}
}
\value{
Returns a table with shares
}
\description{
This function will take in a data set (of class data.table) and will calculate shares
by a specified volume, for a main variable (default is ToolName), and optional groupings.
}
\examples{
##
## Basic use of FnMix
##
library(data.table)
library(easydata)

## Load example dataset:
##  - Then change all potential numeric columns, to numeric
##    classes (this does it safely, quickly, and by reference)
#
data("fiberCountyDem")
NumMorph(fiberCountyDem)

## Calculate each state's share of total US population,
## by each year in our data
#
result <- fnWeight(fiberCountyDem, "State", "Year", sum(Population))
head(result)
}

